+++ Examples/GLImage Demo/GLViewExample/GLViewDrawingExampleController.h
@@ -1,0 +1,16 @@
+//
+//  GLViewDrawingExampleController.h
+//  GLImageDemo
+//
+//  Created by Nick Lockwood on 01/06/2012.
+//
+//
+
+#import <UIKit/UIKit.h>
+#import "GLImageView.h"
+
+@interface GLViewDrawingExampleController : UIViewController
+
+@property (nonatomic, strong) IBOutlet GLImageView *drawingView;
+
+@end