+++ GLView/Core/GLARCHelper.h
@@ -1,7 +1,7 @@
 //
 //  ARC Helper
 //
-//  Version 1.3
+//  Version 1.3.1
 //
 //  Created by Nick Lockwood on 05/01/2012.
 //  Copyright 2012 Charcoal Design